Moved JSON and Avro support behind feature flags.
